Cast as size_t instead of int in clamp usage

+++ src/win32/console.cpp
@@ -271,7 +271,7 @@ void ConsoleModel::deleteHistorySearchChars(size_t n)
 
 void ConsoleModel::redrawInputLine(int p)
 {
-    insertPos = std::clamp<int>(p, 0, buffer.size());
+    insertPos = std::clamp<size_t>(p, 0, buffer.size());
     redrawInputLineNeeded = true;
 }
 
@@ -320,7 +320,7 @@ static bool isWordBoundary(size_t i, const std::wstring s)
 
 int ConsoleModel::detectWordBoundary(int start, bool forward)
 {
-    start = std::clamp<int>(start, 0, buffer.size());
+    start = std::clamp<size_t>(start, 0, buffer.size());
     do
     {
         start += (forward ? 1 : -1);
@@ -330,8 +330,8 @@ int ConsoleModel::detectWordBoundary(int start, bool forward)
 
 void ConsoleModel::deleteCharRange(int start, int end)
 {
-    start = std::clamp<int>(start, 0, buffer.size());
-    end = std::clamp<int>(end, 0, buffer.size());
+    start = std::clamp<size_t>(start, 0, buffer.size());
+    end = std::clamp<size_t>(end, 0, buffer.size());
     if (start < end)
     {
         buffer.erase(start, end - start);
